+++ packages/react/src/components/qr-code/qr-code-context.tsx
@@ -1,0 +1,8 @@
+import type { ReactNode } from 'react'
+import { type UseQrCodeContext, useQrCodeContext } from './use-qr-code-context'
+
+export interface QrCodeContextProps {
+  children: (context: UseQrCodeContext) => ReactNode
+}
+
+export const QrCodeContext = (props: QrCodeContextProps) => props.children(useQrCodeContext())