+++ packages/react/src/components/qr-code/tests/qr-code.test.tsx
@@ -1,0 +1,21 @@
+import { qrCodeAnatomy } from '@ark-ui/anatomy'
+import { cleanup, render } from '@testing-library/react/pure'
+import { QrCode } from '../'
+import { getExports, getParts } from '../../../setup-test'
+import { ComponentUnderTest } from './basic'
+
+describe('Avatar / Parts & Exports', () => {
+  afterAll(() => {
+    cleanup()
+  })
+
+  render(<ComponentUnderTest />)
+
+  it.each(getParts(qrCodeAnatomy))('should render part %s', async (part) => {
+    expect(document.querySelector(part)).toBeInTheDocument()
+  })
+
+  it.each(getExports(qrCodeAnatomy))('should export %s', async (part) => {
+    expect(QrCode[part]).toBeDefined()
+  })
+})