+++ packages/solid/src/components/qr-code/qr-code-pattern.tsx
@@ -1,0 +1,12 @@
+import { mergeProps } from '@zag-js/solid'
+import { type HTMLArkProps, ark } from '../factory'
+import { useQrCodeContext } from './use-qr-code-context'
+
+export interface QrCodePatternProps extends HTMLArkProps<'path'> {}
+
+export const QrCodePattern = (props: QrCodePatternProps) => {
+  const qrCode = useQrCodeContext()
+  const mergedProps = mergeProps(() => qrCode().getPatternProps(), props)
+
+  return <ark.path {...mergedProps} />
+}