+++ packages/vue/src/components/qr-code/examples/basic.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import { QrCode } from '../..'
+</script>
+
+<template>
+  <QrCode.Root value="http://ark-ui.com">
+    <QrCode.Frame>
+      <QrCode.Pattern />
+    </QrCode.Frame>
+    <QrCode.Overlay>
+      <img src="https://ark-ui.com/icon-192.png" alt="" />
+    </QrCode.Overlay>
+  </QrCode.Root>
+</template>