+++ packages/vue/src/components/qr-code/use-qr-code-context.ts
@@ -1,0 +1,6 @@
+import { createContext } from '../../utils'
+import type { UseQrCodeReturn } from './use-qr-code'
+
+export interface UseQrCodeContext extends UseQrCodeReturn {}
+
+export const [QrCodeProvider, useQrCodeContext] = createContext<UseQrCodeContext>('QrCodeContext')