+++ packages/react/CHANGELOG.md
@@ -9,15 +9,14 @@ description: All notable changes will be documented in this file.
 ### Added
 
 - **Signature Pad:** Introduced the `SignaturePad` component for capturing signatures.
+- **QR Code:** Added the `QRCode` component for generating QR codes.
 - **CheckboxGroup:** Added the `CheckboxGroup` component for managing multiple checkboxes.
 
 ### Fixed
 
 - **TreeView:** Resolved an issue where the tree view state could be updated after setting
   `defaultSelectedValue` or `defaultExpandedValue`.
-
 - **Popover, HoverCard:** Fixed issue where `defaultOpen` does not work
-
 - **Tabs:** Resolved an issue where rapidly changing the selected tab could show previous tab
   content briefly.